requirements.txt
alembic==1.17.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
pyotp==2.9.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
PyMuPDF==1.26.7
pytesseract==0.3.13
python-dotenv==1.1.1
python-multipart==0.0.20
qrcode[pil]==7.4.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0

# Chatbot (RAG assistant)
langchain==0.3.27
langchain-community==0.3.27
langchain-groq==0.3.4
langchain-huggingface==0.3.0
faiss-cpu==1.9.0
sentence-transformers==3.3.1
pypdf==5.1.0
reportlab==4.4.10

# Categorization pipeline (LLM fallback)
google-generativeai==0.8.4
groq>=0.28.0

